Added manual data entry and CSV support
Comparison and Program Improvements
Previous Version
External File Dependency:

The previous version relied on an external .txt file to load the data and generate the graphs. Users had to manually prepare a text file with the data, which was necessary for the program to function.

Data Structure:

The program required the data to be organized in a .txt file with the first column dedicated to times and subsequent columns to temperature values.
Limited Features:

The program allowed changing the names of the columns and viewing the data loaded from the .txt file.
Chart interaction capabilities were limited to basic functionality such as zooming, moving the chart, and restoring the original view.
The graphical interface lacked advanced options to enter data manually or to process files in other formats such as CSV.

Chart Customization:

Users could add reference lines to the graph to show the 2°C and 8°C temperature limits, but they could not customize these lines in more detail.
Improved Version
Manual Data Integration:

Added functionality that allows users to manually enter data directly through the graphical interface, eliminating the need to upload an external .txt file. This makes the application more versatile and accessible.
Flexibility in Data Structure:

Removed the restriction that the first column of the .txt file had to be dedicated to time. Now, any column can represent any type of data, allowing for greater flexibility in data entry.

Expansion of Functionalities:

Added the ability to select and process up to five CSV files, extracting and combining the data into a new .txt file. This allows the integration of data from multiple sources and formats.
New advanced options such as dynamic dashed line settings, improved zoom and chart movement, and customizing column names from the interface.
Personalization and Advanced Settings:

Users can now configure specific values ​​for the dashed lines that represent temperature limits on the graph, providing greater control over the display of data.
Introduced the option to change data column names and save them to a JSON file, allowing for greater customization and reuse of configurations.